Polished the code to prepare for a code review.
I added a constant to the componentConfig called MAX_BACKGROUND_MESSAGES that I used to make sure that the data collected message only runs for the first few iterations to keep the event panel cleaner.

I also added a variable called VIRTUAL that if defined in the config file will run the dummy EPS and disable the code that you added to run the camera and the imu.

I also renamed one of the variables in the sendTransmission command so that it better matched up with what was actually happening behind the scenes.