Cleaned the code and fixed the imu so that it is officially working and working well.
Added documentation and comments to IMU.

Fixed the IMU so that it writes correctly and reverted how it reads so that reading also works.

Added a picture to the Flight Logic documentation.

Removed a lot of ports connected to the data collector that were originally used only as a proof of concept.