Started a new version of the FSWDeployment for FlatSat 2.0
Deleted the old FSW Deployment

Created a new one that implements the watchDog component and the LinuxInterrupt

Created a dummy folder to hold dummy drivers

Created a dummy GPIO driver to GPIO stuff locally

Tested the system locally

Updated the deployment to use a real GPIO port but didn't have time to test. Untested on hardware.